promote message_v2 module to message
