+++ cli/src/main.rs
@@ -9,7 +9,7 @@ use clap::{
 use git_version::git_version;
 
 
-use printnanny_nats::message_v2::{NatsReply, NatsRequest};
+use printnanny_nats::message::{NatsReply, NatsRequest};
 use printnanny_nats::cloud_worker::DEFAULT_NATS_CLOUD_APP_NAME;
 use printnanny_nats::subscriber::{ NatsSubscriber, DEFAULT_NATS_EDGE_APP_NAME};
 use printnanny_services::settings::SettingsFormat;
