+++ nats/src/bin/nats-edge-worker.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use printnanny_nats::message_v2::{NatsReply, NatsRequest};
+use printnanny_nats::message::{NatsReply, NatsRequest};
 use printnanny_nats::subscriber::NatsSubscriber;
 
 #[tokio::main]
