+++ nats/src/lib.rs
@@ -2,7 +2,7 @@ pub mod cloud_commands;
 pub mod cloud_publisher;
 pub mod cloud_worker;
 // pub mod message;
-pub mod message_v2;
+pub mod message;
 pub mod subjects;
 pub mod subscriber;
 pub mod util;
